+++ Rectangles/Rectangles/XYPoint.h
@@ -12,5 +12,5 @@
 
 @property int x, y;
 
--(void) setX: (int) xVal andY: (int) yVal; 
+-(void) setX: (float) xVal andY: (float) yVal; 
 @end
